gtkwave/pyproject.toml
[tool.poetry]
name = "pio2fst"
version = "0.1.0"
description = "Visualisation of PIO program using GTKWave"
authors = ["Nathan Young"]
license = "Apache-2.0"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
adafruit-circuitpython-pioasm = "1.0.0"
click = "8.1.7"
pylibfst = "0.2.0"
rp2040-pio-emulator = "0.83.0"

[tool.poetry.group.dev.dependencies]
black = "24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pio2fst = "pio2fst.__main__:run"
